social network

+++ yatube/posts/views.py
@@ -4,8 +4,8 @@ from django.contrib.auth.decorators import login_required
 
 from yatube.settings import PAGE_COUNT
 
-from .models import Group, Post, User, Follow
-from .forms import PostForm, CommentForm
+from posts.models import Group, Post, User, Follow
+from posts.forms import PostForm, CommentForm
 
 
 from django.core.paginator import Paginator
@@ -50,14 +50,10 @@ def profile(request, username):
     paginator = Paginator(posts, PAGE_COUNT)
     page_number = request.GET.get("page")
     page_obj = paginator.get_page(page_number)
-    subscribers = post.following.all()
     following = False
-    authors = list()
-    for user in subscribers:
-        author = user.user
-        authors.append(author)
-    if request.user in authors:
+    if request.user in post.following.all():
         following = True
+
     context = {
         "post": post,
         "page_obj": page_obj,
@@ -117,23 +113,18 @@ def post_edit(request, post_id):
 def add_comment(request, post_id):
     post = get_object_or_404(Post, id=post_id)
     form = CommentForm(request.POST or None)
-    if form.is_valid():
-        comment = form.save(commit=False)
-        comment.author = request.user
-        comment.post = post
-        comment.save()
+    if not form.is_valid():
+        return redirect("posts:post_detail", post_id)
+    comment = form.save(commit=False)
+    comment.author = request.user
+    comment.post = post
+    comment.save()
     return redirect("posts:post_detail", post_id)
 
 
 @login_required
 def follow_index(request):
-    user = request.user
-    followings = user.follower.all()
-    authors = list()
-    for user in followings:
-        author = user.author
-        authors.append(author)
-    post_list = Post.objects.filter(author__in=authors)
+    post_list = Post.objects.filter(author__following__user=request.user)
     paginator = Paginator(post_list, PAGE_COUNT)
     page_number = request.GET.get("page")
     page_obj = paginator.get_page(page_number)
@@ -146,12 +137,8 @@ def follow_index(request):
 @login_required
 def profile_follow(request, username):
     post_author = get_object_or_404(User, username=username)
-    subscribers = post_author.following.all()
-    authors = list()
-    for user in subscribers:
-        author = user.user
-        authors.append(author)
-    if request.user != post_author and request.user not in authors:
+    if (request.user != post_author
+            and request.user != post_author.following.all()):
         Follow.objects.create(user=request.user, author=post_author)
     return redirect("posts:profile", username=username)
 
@@ -159,5 +146,6 @@ def profile_follow(request, username):
 @login_required
 def profile_unfollow(request, username):
     post_author = get_object_or_404(User, username=username)
-    Follow.objects.filter(user=request.user, author=post_author).delete()
+    if Follow.objects.filter(user=request.user, author=post_author).exists():
+        Follow.objects.filter(user=request.user, author=post_author).delete()
     return redirect("posts:profile", username=username)
